Combine the modules and update the binaries
